+++ src/webapp/__init__.py
@@ -3,8 +3,8 @@ from random import randint
 from time import strftime
 
 from flask import Flask, flash, jsonify, render_template, request
-from flaskr.data_src import DataStructures
-from flaskr.mealplan import MealplanGenerator
+from webapp.data_src import DataStructures
+from webapp.mealplan import MealplanGenerator
 from wtforms import Form, StringField, SubmitField, validators
 
 app = Flask(__name__)
