+++ src/webapp/templates/exerciseplan.html
