+++ src/webapp/templates/exercises.html
