+++ src/webapp/templates/points.html
