+++ src/webapp/templates/shoppinglist.html
